Include Wikibase Extensions & split components
This creates a new page for Wikibase extensions.
The page is split into:
 - Repo Extensions
 - Client Extensions
 - WMF specific extensions

This also split the components page
(and actually gets rid of it) choosing to put the libraries
that are components of wikibase on the applications page
and other libraries on their own page.

This will likely lead to people being able to find the
client libraries that they want a bit quicker, rather than
first having to sift through the wikibase components they dont
care about.

Change-Id: I05d3ddcfaad328edfa6f027eacbffa1184877dce